+++ src/app/(route)/hello/_components/ImageSlots/ThirdSection/ThirdSection.stories.tsx
