--- src/app/(route)/hello/_hooks/index.ts
@@ -1,1 +1,0 @@
-export { useInView } from "./useInView/useInView";
