+++ src/app/(route)/list/[id]/_hooks/useToggleFavorite/useToggleFavorite.ts
